msm8998-common: start RIL for second SIM later
Sometimes, at boot, only the second SIM is recognized by Android.
It turned out that this happens only when the RIL for the SIM1
(service vendor.ril-daemon) starts after the RIL for the SIM2
(service vendor.ril-daemon2).

To solve this, simply start the RIL for the SIM2 later in the boot
process: to be sure wait for the boot to be completed, that is a
bit late, but it works fine.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>