msm8998-common: start RIL for second SIM later [2]
While starting the RIL for the second SIM later works usually fine,
there are cases where it doesn't work: for example, after a factory
reset or an OTA upgrade it appears that restart the service is also
needed.
All of this takes a bit (~10seconds) after the lockscreen is shown after
the boot is completed, but at least both SIMs should always be detected
properly.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>